docker build
Eventually I’d like to go the route of having everything in this repo
managed via Docker so people don’t have to deal with deps, so might as
well start here.

This just adds a working Dockerfile build, but doesn’t transition our
runtime over to using it yet.

If we do go that route eventually, it will hopefully give us a more
consistent way to do the stats portion as well.